+++ CHANGELOG.md
@@ -1,3 +1,6 @@
 ## v0.0.1
 
 * Initial release
+* Main supported fields;
+* Generators;
+* Integration logic.